requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dateutil==2.8.2
schedule==1.2.0
requests==2.31.0
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0gunicorn==21.2.0
